More splitting of lux metering, plus fix existing bugs.
