+++ vdu_controls/vdu_controls_main.py
@@ -1,3 +1,5 @@
+# SPDX-FileCopyrightText: 2021-2026 Contributors to vdu_controls <https://github.com/digitaltrails/vdu_controls>
+# SPDX-License-Identifier: GPL-3.0-or-later
 import sys
 
 import vdu_controls.vdu_controls_application as application
